+++ src/main/java/xyz/nikitacartes/easyauth/commands/LogoutCommand.java
@@ -2,6 +2,7 @@ package xyz.nikitacartes.easyauth.commands;
 
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
+import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.minecraft.network.packet.s2c.play.PlayerListS2CPacket;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -16,6 +17,7 @@ public class LogoutCommand {
     public static void registerCommand(CommandDispatcher<ServerCommandSource> dispatcher) {
         // Registering the "/logout" command
         dispatcher.register(literal("logout")
+                .requires(Permissions.require("easyauth.commands.logout", true))
                 .executes(ctx -> logout(ctx.getSource())) // Tries to de-authenticate the user
         );
     }